Move TestSampleClass into own file
